🐛 Fix : 제미나이 리뷰 반영..3
