+++ src/app/(service)/boards/utils/boardDisplayUtils.ts
@@ -2,6 +2,8 @@
  * 게시글 표시 유틸리티 함수
  */
 
+import { formatRelativeTime } from '@/utils/formatDate';
+
 export const getLikeCount = (likeCount: number) => {
   return likeCount > 999 ? '999+' : likeCount.toString();
 };
@@ -17,25 +19,18 @@ export const formatDateToYmd = (value: string) => {
   const month = String(date.getMonth() + 1).padStart(2, '0');
   const day = String(date.getDate()).padStart(2, '0');
 
-  return `${year}.${month}.${day}`;
+  return `${year}-${month}-${day}`;
 };
 
 export const formatRelativeOrYmdHm = (value: string) => {
-  const date = new Date(value);
-
-  if (Number.isNaN(date.getTime())) {
-    return value;
-  }
-
-  const diffMs = Date.now() - date.getTime();
-  const diffMinutes = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-
-  if (diffMinutes < 1) return '방금 전';
-  if (diffHours < 1) return `${diffMinutes}분 전`;
-  if (diffHours < 24) return `${diffHours}시간 전`;
-
-  return formatDateToYmdHm(value);
+  return formatRelativeTime(value, (date) => {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0');
+    const day = String(date.getDate()).padStart(2, '0');
+    const hours = String(date.getHours()).padStart(2, '0');
+    const minutes = String(date.getMinutes()).padStart(2, '0');
+    return `${year}-${month}-${day} ${hours}:${minutes}`;
+  });
 };
 
 export const formatDateToYmdHm = (value: string) => {
