+++ src/utils/formatDate.ts
@@ -18,3 +18,31 @@ export function formatKSTTime(dateString?: string): string | null {
   if (Number.isNaN(date.getTime())) return null;
   return KST_TIME_FORMAT.format(date);
 }
+
+/**
+ * 24시간 이내: '방금 전' / 'n분 전' / 'n시간 전'
+ * 24시간 초과: formatFallback이 있으면 그 결과, 없으면 'yyyy-mm-dd'
+ */
+export function formatRelativeTime(
+  value: string | null,
+  formatFallback?: (date: Date) => string,
+): string {
+  if (!value) return '';
+  const date = new Date(value);
+  if (Number.isNaN(date.getTime())) return '';
+
+  const diffMs = Date.now() - date.getTime();
+  const diffMinutes = Math.floor(diffMs / 60000);
+  const diffHours = Math.floor(diffMs / 3600000);
+
+  if (diffMinutes < 1) return '방금 전';
+  if (diffHours < 1) return `${diffMinutes}분 전`;
+  if (diffHours < 24) return `${diffHours}시간 전`;
+
+  if (formatFallback) return formatFallback(date);
+
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}-${month}-${day}`;
+}
